+++ src/main/java/com/boanni_back/project/config/SecurityConfig.java
@@ -41,23 +41,20 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
-            .headers(headers -> headers.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin))
-            .cors(cors->cors.configurationSource(corsConfigurationSource()))
-            .formLogin(AbstractHttpConfigurer::disable)
-            .httpBasic(AbstractHttpConfigurer::disable)
-        // csrf 인증 비활성화, 나중에 비활성화 코드 지우기
-            .csrf(AbstractHttpConfigurer::disable)
-            .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-            .authorizeHttpRequests(auth -> auth
+                .headers(headers -> headers.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin))
+                .cors(cors->cors.configurationSource(corsConfigurationSource()))
+                .formLogin(AbstractHttpConfigurer::disable)
+                .httpBasic(AbstractHttpConfigurer::disable)
+                // csrf 인증 비활성화, 나중에 비활성화 코드 지우기
+                .csrf(AbstractHttpConfigurer::disable)
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .authorizeHttpRequests(auth -> auth
 //                        해당 경로는 모두 접근 가능하다.
-                    .requestMatchers("/auth/verify**","/auth/signin", "/auth/signup", "/h2-console/**").permitAll()
-                    .requestMatchers("/admin/**", "/api/users/**", "/api/questions/**", "/images/**", "/api/record/**").permitAll()
-                    .requestMatchers("/auth/**", "/h2-console/**").permitAll()
-//                  이외 요청은 jwt 토큰이 없으면 접근 불가능하다.
-                    .anyRequest().authenticated()
-                    //.anyRequest().permitAll() //개발 중에만 모두 허용
-            )
-            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);  // 🔥 조립
+                                .requestMatchers("/auth/**","/admin/**", "/h2-console/**", "/questions/admin/**").permitAll()
+////                        이외 요청은 jwt 토큰이 없으면 접근 불가능하다.
+                                .anyRequest().authenticated()
+                )
+                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);  // 🔥 조립
 
         return http.build();
     }
